gupiao3/requirements.txt
akshare>=1.12.0
pandas>=2.0.0
matplotlib>=3.7.0
requests>=2.31.0
pyyaml>=6.0             # 用于读取配置文件